refactor(Member VO) : MemberName, MemberProfileImageKey 등의 VO 클래스를 domain.model.vo 패키지로 이동
